Separate SynthEventHandler into two classes
