+++ lib/src/processor/SynthProcessorCore.ts
@@ -2,6 +2,7 @@ import { SampleParameter, SampleRange, SynthEvent } from "../SynthEvent"
 import { logger } from "./logger"
 import { Sample, SampleTable } from "./SampleTable"
 import { SynthEventHandler } from "./SynthEventHandler"
+import { SynthEventScheduler } from "./SynthEventScheduler"
 import { WavetableOscillator } from "./WavetableOscillator"
 
 interface ChannelState {
@@ -36,12 +37,18 @@ const RHYTHM_BANK = 128
 export class SynthProcessorCore {
   private sampleTable = new SampleTable()
   private channels: { [key: number]: ChannelState } = {}
-  private readonly eventHandler: SynthEventHandler
-  private readonly sampleRate: number
-  private readonly getCurrentFrame: () => number
+  private readonly eventScheduler: SynthEventScheduler
 
-  constructor(sampleRate: number, getCurrentFrame: () => number) {
-    this.eventHandler = new SynthEventHandler(this)
+  constructor(
+    private readonly sampleRate: number,
+    private readonly getCurrentFrame: () => number
+  ) {
+    const eventHandler = new SynthEventHandler(this)
+    this.eventScheduler = new SynthEventScheduler(
+      getCurrentFrame,
+      (e) => eventHandler.handleImmediateEvent(e),
+      (e) => eventHandler.handleDelayableEvent(e)
+    )
     this.sampleRate = sampleRate
     this.getCurrentFrame = getCurrentFrame
   }
@@ -70,7 +77,7 @@ export class SynthProcessorCore {
   }
 
   addEvent(e: SynthEvent) {
-    this.eventHandler.addEvent(e)
+    this.eventScheduler.addEvent(e)
   }
 
   noteOn(channel: number, pitch: number, velocity: number) {
@@ -153,6 +160,7 @@ export class SynthProcessorCore {
   }
 
   allSoundsOff(channel: number) {
+    this.eventScheduler.removeScheduledEvents(channel)
     const state = this.getChannelState(channel)
 
     for (const key in state.oscillators) {
@@ -220,7 +228,7 @@ export class SynthProcessorCore {
   }
 
   process(outputs: Float32Array[]): void {
-    this.eventHandler.processScheduledEvents()
+    this.eventScheduler.processScheduledEvents()
 
     for (const channel in this.channels) {
       const state = this.channels[channel]
